+++ src/main/java/com/github/mjkuranda/spaceadventure2/states/IntroState.java
@@ -1,5 +1,6 @@
 package com.github.mjkuranda.spaceadventure2.states;
 
+import com.github.mjkuranda.spaceadventure2.PlayerStatistics;
 import org.newdawn.slick.*;
 import org.newdawn.slick.state.BasicGameState;
 import org.newdawn.slick.state.StateBasedGame;
@@ -32,6 +33,7 @@ public class IntroState extends BasicGameState {
 
         if (in.isKeyPressed(Input.KEY_1)) {
             game.enterState(StatesId.GAME, new FadeOutTransition(), new FadeInTransition());
+            PlayerStatistics.getInstance().clear();
         }
     }
 }